+++ src/main/kotlin/csense/kotlin/not/highlighter/bll/TextHighlightDecider.kt
@@ -1,6 +1,6 @@
 package csense.kotlin.not.highlighter.bll
 
-import com.intellij.openapi.util.TextRange
+import com.intellij.openapi.util.*
 import csense.idea.base.bll.openApi.*
 import csense.kotlin.extensions.primitives.*
 import csense.kotlin.not.highlighter.settings.*
@@ -33,20 +33,6 @@ class TextHighlightDecider(
     }
 
     companion object {
-        const val exclMarkText: String = "!"
-        const val notText: String = "not"
-        const val disableText: String = "disable"
-        const val disabledText: String = "disabled"
+
     }
 }
-
-fun TextHighlightDecider.Companion.disabledTextsOrEmpty(fromSettings: NotHighlighterSettings): List<String> {
-    if (!fromSettings.highlightDisabledText) {
-        return emptyList()
-    }
-    return listOf(
-        disableText,
-        disabledText
-    )
-}
-