+++ src/main/kotlin/csense/kotlin/not/highlighter/highligter/AnnotationHolderHighlighterStrategy.kt
@@ -1,26 +1,34 @@
 package csense.kotlin.not.highlighter.highligter
 
 import com.intellij.lang.annotation.*
-import com.intellij.openapi.editor.markup.*
 import com.intellij.openapi.util.*
 import com.intellij.psi.*
 import csense.idea.base.bll.*
 import csense.idea.base.bll.string.*
 import csense.kotlin.not.highlighter.bll.*
+import csense.kotlin.not.highlighter.highligter.elementHighligter.settings.*
 import org.jetbrains.kotlin.psi.*
 
 class AnnotationHolderHighlighterStrategy(
-    private val textHighlightDecider: TextHighlightDecider,
-    private val textAttributes: TextAttributes
+    private val textHighlightDecider: TextHighlightDecider
 ) {
-    fun highlightNameIdentifier(element: KtNamedDeclaration, holder: AnnotationHolder) {
+    fun highlightNameIdentifier(
+        element: KtNamedDeclaration,
+        holder: AnnotationHolder,
+        setting: HighlighterSetting
+    ) {
         val nameIdentifier: PsiElement = element.nameIdentifier ?: return
-        highlightTextIn(element = nameIdentifier, holder = holder)
+        highlightTextIn(
+            element = nameIdentifier,
+            holder = holder,
+            setting = setting
+        )
     }
 
     fun highlightTextIn(
         element: PsiElement,
-        holder: AnnotationHolder
+        holder: AnnotationHolder,
+        setting: HighlighterSetting
     ) {
         element.camelCase.forEachCamelCaseWordWithTextRange { originalRange: TextRange, string: String ->
             if (textHighlightDecider.shouldNotHighlight(string)) {
@@ -34,12 +42,20 @@ class AnnotationHolderHighlighterStrategy(
 
             highlightRange(
                 range = rangeToHighlight,
-                holder = holder
+                holder = holder,
+                setting = setting
             )
         }
     }
 
-    fun highlightRange(range: TextRange, holder: AnnotationHolder) {
-        holder.highlightTextRange(range = range, withStyle = textAttributes)
+    fun highlightRange(
+        range: TextRange,
+        holder: AnnotationHolder,
+        setting: HighlighterSetting
+    ) {
+        holder.highlightTextRange(
+            range = range,
+            withStyle = setting.textAttributes()
+        )
     }
 }