+++ src/main/kotlin/csense/kotlin/not/highlighter/highligter/elementHighligter/highlighterStrategy/DisabledHighlighter.kt
@@ -1,7 +1,8 @@
-package csense.kotlin.not.highlighter.highligter.elementHighligter
+package csense.kotlin.not.highlighter.highligter.elementHighligter.highlighterStrategy
 
 import com.intellij.lang.annotation.*
 import csense.kotlin.not.highlighter.bll.*
+import csense.kotlin.not.highlighter.highligter.elementHighligter.*
 
 object DisabledHighlighter : ElementHighlighter {
     override fun mayHighlight(): Boolean {