+++ changelog.md
@@ -1,3 +1,8 @@
+# 2.1.0
+- can now set color independently for operators and function name(s) and comments
+- added ability to add custom names to highlight
+
+
 # 2.0.0
 - Now supports 'not' & '!' highlighting in strings and comments.
 - Also highlights "Disabled" by default (can be toggled off in settings page)